Fix test imports after moving to __tests__ folder
